+++ src/Shared/ResourceSemanticConventions.cs
@@ -49,7 +49,6 @@ internal static class ResourceSemanticConventions
     public const string AttributeCloudProvider = "cloud.provider";
     public const string AttributeCloudRegion = "cloud.region";
     public const string AttributeCloudResourceId = "cloud.resource_id";
-    public const string AttributeCloudZone = "cloud.zone";
     public const string AttributeComponent = "component";
 
     public const string AttributeOsType = "os.type";