+++ src/OpenTelemetry.Resources.Gcp/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Updated OpenTelemetry core component version(s) to `1.17.0`.
+  ([#4773](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4773))
+
 ## 1.0.0-alpha.2
 
 Released 2026-Jul-09